+++ HISTORY.md
@@ -1,8 +1,13 @@
 History of features added to LiChess Tools in time (reversed order)
 
+08 Apr
+
+- Toggle practice quick action
+
 07 Apr
 
 - custom engine level for practice
+- full data storage for local eval cache
 
 22 Mar